Style pages based on the collection, not section
We were previously using the `section` variable to denote the current
guide being shown, but this wasn't a good use of the nomenclature.

We will now refer to guides by looking at the `collection` variable, 
which is set automatically by Jekyll. We still include the `section` 
variable in the `<body>` class to support the `home` value, for now.