Automatically add previous and next page links
This adds a Liquid include to automatically build links to the previous
and next pages of each guide.

Due to Jekyll quirks, and because we've added sections, it's a bit 
hacky - I encourage you to read the inline Liquid comments of 
`_includes/previous_next.html` to understand how the correct page is
chosen.